Refactored RecordingMap into own file
