+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import 'package:stasi/db/database_bloc.dart';
-import 'package:stasi/pages/running_recording.dart';
+import 'package:stasi/notifiers/running_recording.dart';
 import 'package:stasi/util/theme.dart';
 import 'package:stasi/pages/recording_manager.dart';
 import 'package:stasi/pages/vehicle_selection.dart';
